// src/api/music.ts
export interface Song {
  id: number;
  title: string;
  artist: Artist;
  album: Album;
  digest: string;
  filename: string;
  created_at: string;
  time: number;
  year: number | null;
  rating: number;
  played_count: number;
  weight?: number;
}

export interface Metadata {
  format: {
    codec: string;
    length: number;
    bitrate: string;
    sample_rate: string;
    channels: string;
  };
  tags: {
    title: string | null;
    artist: string | null;
    album: string | null;
    album_artist: string | null;
    track: string | null;
    disc: string | null;
    year: string | null;
    lyrics: string | null;
    cover_art?: {
      mime: string;
      data: string; // base64
    }
  };
}

export interface Album {
  id: number;
  title: string;
  artist: string;
}

export interface Artist {
  id: number;
  name: string;
  ruby: string;
}

export interface Playlist {
  id: number;
  name: string;
  songs?: Song[];
}